Support de tout type de fichiers (#4)
* chore: added .gitattributes for .sh files

This is to prevent Git from changing EOL sequence to CRLF on Windows, causing apt to install nginx\r instead of nginx, leading to an error

* chore(deps): bump all dependencies

* chore: renamed /images directory to /files to be more generic

* fix: metrics route hardcoded /images directory instead of using the one in the settings

* fix: double dot in the cache directory containing resized images

* feat: added support for all files type

This change added two new settings. The first one is a list of MIME types that are resizable (on which the w and h query parameter will apply) and the second one is the list of allowed MIME file types to be uploaded. If an user tries to upload a file, and the MIME type of this file isn't in the allowed MIME types list, an error will be raised

* docs: added explanation on settings values

* feat: metrics are now grouped per mime type

* chore: improved Dockerfile

* docs: added section inside settings

* ci: bump version of python base docker image

* chore: added back empty last lines